requirements.txt
smolagents[litellm]
huggingface_hub
requests
python-dotenv
